librados: Added new API methods to create an ioctx by pool id
A new pool_list method includes the unique pool id for all pools.
This pool id can be used with the new rados_ioctx_create2 /
ioctx_create methods to create an ioctx by pool id instead of
by pool name.  Creating ioctx's by pool id helps avoid certain
race conditions when pools are renamed or deleted.

+++ src/librados/IoCtxImpl.cc
@@ -33,7 +33,7 @@ librados::IoCtxImpl::IoCtxImpl() :
 }
 
 librados::IoCtxImpl::IoCtxImpl(RadosClient *c, Objecter *objecter,
-			       Mutex *client_lock, int poolid,
+			       Mutex *client_lock, int64_t poolid,
 			       const char *pool_name, snapid_t s)
   : ref_cnt(0), client(c), poolid(poolid), pool_name(pool_name), snap_seq(s),
     assert_ver(0), notify_timeout(c->cct->_conf->client_notify_timeout),
